move strategies to proper places
