+++ nle_code_wrapper/bot/strategies/__init__.py
@@ -3,33 +3,32 @@ from nle_code_wrapper.bot.strategies.explore import (
     explore_room,
     explore_room_systematically,
     general_explore,
-    goto_next_unexplored_room,
-    search_corridor_for_hidden_doors,
-    search_for_traps,
-    search_room_for_hidden_doors,
 )
 from nle_code_wrapper.bot.strategies.explore_items import explore_items
 from nle_code_wrapper.bot.strategies.fight_monster import fight_all_monsters, fight_closest_monster
-from nle_code_wrapper.bot.strategies.goto import goto
-from nle_code_wrapper.bot.strategies.goto_items import goto_items
-from nle_code_wrapper.bot.strategies.goto_stairs import goto_stairs
+from nle_code_wrapper.bot.strategies.goto import goto, goto_closest_unexplored_room, goto_items, goto_stairs
 from nle_code_wrapper.bot.strategies.open_doors import open_doors, open_doors_key, open_doors_kick
 from nle_code_wrapper.bot.strategies.pickup_items import pickup_items
 from nle_code_wrapper.bot.strategies.quaff import quaff_potion_from_inv
 from nle_code_wrapper.bot.strategies.random_move import random_move
 from nle_code_wrapper.bot.strategies.run_away import run_away
-from nle_code_wrapper.bot.strategies.search import search
+from nle_code_wrapper.bot.strategies.search import (
+    general_search,
+    search_corridor_for_hidden_doors,
+    search_for_traps,
+    search_room_for_hidden_doors,
+)
 
 __all__ = [
     explore_corridor,
     explore_room,
     explore_room_systematically,
     general_explore,
-    goto_next_unexplored_room,
+    goto_closest_unexplored_room,
     search_corridor_for_hidden_doors,
     search_for_traps,
     search_room_for_hidden_doors,
-    search,
+    general_search,
     explore_items,
     goto_stairs,
     fight_closest_monster,
