--- nle_code_wrapper/bot/strategies/goto_stairs.py
@@ -1,25 +1,0 @@
-from nle_utils.glyph import G
-
-from nle_code_wrapper.bot import Bot
-
-
-def goto_stairs(bot: "Bot") -> bool:
-    """
-    Directs the bot to move towards the stairs on the current level.
-    This function attempts to find the coordinates of the stairs down on the current level
-    and directs the bot to move towards them using the bot's pathfinder. If a path to
-    the stairs is found, the bot will move towards the first set of stairs found.
-    Args:
-        bot (Bot): The bot instance that will be directed to the stairs.
-    Returns:
-        bool: True if the bot successfully finds a path to the stairs and starts moving
-              towards them, False otherwise.
-    """
-
-    level = bot.current_level
-    stairs = level.object_coords(G.STAIR_DOWN)
-    if len([s for s in stairs if bot.pathfinder.get_path_to(s)]) > 0:
-        bot.pathfinder.goto(stairs[0])
-        return True
-    else:
-        return False
